+++ nixos/modules/profile/usage/workstation.nix
@@ -69,15 +69,6 @@ in
     ];
 
     ########################################################################
-    ##                           Kernel modules                           ##
-    ########################################################################
-
-    boot = mkIf (release == "20.09") {
-      extraModulePackages = with config.boot.kernelPackages; [ exfat-nofuse ];
-      kernelModules = [ "exfat" ];
-    };
-
-    ########################################################################
     ##                              Hardware                              ##
     ########################################################################
 
@@ -113,16 +104,13 @@ in
         # Configure the keyboard layout if it has been set in
         # confkit.keyboard.layout.
         layout = mkIf (layout == "bépo") "fr";
-        xkbVariant = mkIf (layout == "bépo") "bepo";
+        xkbVariant = mkIf (layout == "bépo") "bepo_afnor";
 
         # Enable touchpad support with natural scrolling.
         libinput = {
           enable = true;
-        } // (if release == "20.09" then {
-          naturalScrolling = true;
-        } else {
           touchpad.naturalScrolling = true;
-        });
+        };
       };
     };